Полностью дописаны алгоритм и эталон сарказма
