+++ engine/stages/main_sarc.py
@@ -1,0 +1,7 @@
+from sarcasm.etalon_sarc import morph_analyzer
+
+
+def main():
+    sarcasm = morph_analyzer()
+    #not_sarcasm = f
+    print(sarcasm)
