+++ engine/stages/sarcasm/algorithm_proj.py
@@ -1,0 +1,28 @@
+import nltk
+
+import pymorphy2
+
+morph = pymorphy2.MorphAnalyzer()
+
+nltk.download("")
+
+def display_counts(count_uppercase,count_unions, count_particle, count_gain_punct, count_interj, timer ):
+    print("\nКоличество букв в верхнем регистре (переменная х1): ", count_uppercase)
+    print("Количество союзов в тексте:", count_unions)
+    print("Количество частиц в тексте:", count_particle)
+    print("Количество усиливающих знаков препинания в тексте:", count_gain_punct)
+    print("Количество междометий в тексте:", count_interj)
+    print(timer)
+async def counters(len_symb_text, list_parts, len_text, f, len_sent, timer):
+
+
+
+
+""""
+standard_100 = open(f'/standards/standard_100', 'w')
+with open(f'/standards/standard_100', 'w'):
+    display_counts(count_uppercase, count_unions,
+                   count_particle, count_gain_punct,
+                    count_interj)
+"""
+
