+++ engine/stages/sarcasm/data.txt
